refactor WorkerManager to use mixins; update unit tests to pass against mixin format (in preparation for testing each mixin)
